refactored parsers, data manager
